Made it more compact, still not neat though
