+++ Bot Scout RAPID REACT/Bot Scout RAPID REACT.csproj
@@ -90,37 +90,19 @@
     <Compile Include="Average.Designer.cs">
       <DependentUpon>Average.cs</DependentUpon>
     </Compile>
-    <Compile Include="Data.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="Data.Designer.cs">
-      <DependentUpon>Data.cs</DependentUpon>
-    </Compile>
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
-    <Compile Include="BestTeams.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="BestTeams.Designer.cs">
-      <DependentUpon>BestTeams.cs</DependentUpon>
-    </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Average.resx">
       <DependentUpon>Average.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Data.resx">
-      <DependentUpon>Data.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="BestTeams.resx">
-      <DependentUpon>BestTeams.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
